# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  domain.LearningPathResponse:
    properties:
      category:
        type: string
      completed_lessons:
        type: integer
      created_at:
        type: string
      description:
        type: string
      duration:
        description: in minutes
        type: integer
      id:
        type: string
      is_enrolled:
        type: boolean
      lesson_ids:
        items:
          type: string
        type: array
      level:
        description: beginner, intermediate, advanced
        type: string
      title:
        type: string
      total_lessons:
        type: integer
      updated_at:
        type: string
      user_progress:
        type: number
    type: object
  domain.LessonResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration:
        description: in minutes
        type: integer
      id:
        type: string
      is_completed:
        type: boolean
      order:
        type: integer
      path_id:
        type: string
      quiz:
        $ref: '#/definitions/domain.Quiz'
      quiz_id:
        type: string
      title:
        type: string
      type:
        description: video, text, quiz, exercise
        type: string
      updated_at:
        type: string
    type: object
...
  domain.ProgressResponse:
    properties:
      completed_lessons:
        items:
          type: string
        type: array
      current_lesson:
        type: string
      enrolled_at:
        type: string
      last_accessed_at:
        type: string
      path_id:
        type: string
      path_title:
        type: string
      progress:
        type: number
      total_lessons:
        type: integer
    type: object
...
  domain.Question:
    properties:
      correct:
        type: string
      id:
        type: string
      options:
        items:
          type: string
        type: array
      points:
        type: integer
      text:
        type: string
      type:
        description: multiple_choice, true_false, text
        type: string
    type: object
  domain.Quiz:
    properties:
      created_at:
        type: string
      id:
        type: string
      lesson_id:
        type: string
      questions:
        items:
          $ref: '#/definitions/domain.Question'
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  domain.QuizResultResponse:
    properties:
      answers:
        additionalProperties:
          type: string
        type: object
      correct_answers:
        additionalProperties:
          type: string
        type: object
      created_at:
        type: string
      lesson_id:
        type: string
      max_score:
        type: integer
      passed:
        type: boolean
      percentage:
        type: number
      quiz_id:
        type: string
      score:
        type: number
    type: object
  domain.QuizSubmissionRequest:
    properties:
      answers:
        additionalProperties:
          type: string
        description: question_id -> answer
        type: object
      quiz_id:
        type: string
    required:
    - answers
    - quiz_id
    type: object
...
  domain.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: AI-powered English coach for Ethiopians seeking global job opportunities
  title: LissanAI API
  version: "1.0"
paths:
  /api/v1/learning/lessons/{id}:
    get:
      consumes:
      - application/json
      description: Fetch the content for a specific lesson
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LessonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get lesson content
      tags:
      - Learning
  /api/v1/learning/lessons/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark a lesson as completed for the authenticated user
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark lesson as completed
      tags:
      - Learning
  /api/v1/learning/paths:
    get:
      consumes:
      - application/json
      description: Retrieve all available learning paths with user progress if enrolled
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.LearningPathResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all learning paths
      tags:
      - Learning
  /api/v1/learning/paths/{id}/enroll:
    post:
      consumes:
      - application/json
      description: Enroll the authenticated user in a specific learning path
      parameters:
      - description: Learning Path ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enroll in a learning path
      tags:
      - Learning
  /api/v1/learning/paths/{id}/progress:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's progress for a specific learning path
      parameters:
      - description: Learning Path ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ProgressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user progress for a learning path
      tags:
      - Learning
  /api/v1/learning/quizzes/{id}/submit:
    post:
      consumes:
      - application/json
      description: Submit user's answers to a quiz for grading
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      - description: Quiz answers
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.QuizSubmissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.QuizResultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit quiz answers
      tags:
      - Learning
  /api/v1/pronunciation/activity:
    post:
      consumes:
      - application/json
      description: Record that a user completed a pronunciation practice session for
        streak tracking
      parameters:
      - description: Pronunciation session data
        in: body
        name: request
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Record pronunciation practice activity
      tags:
      - Pronunciation
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"